+++ services/queue-manage/src/main/java/com/ticketPing/queue_manage/presentaion/controller/WaitingQueueController.java
@@ -15,8 +15,8 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import reactor.core.publisher.Mono;
 
-@RequiredArgsConstructor
 @RestController
+@RequiredArgsConstructor
 @RequestMapping("/api/v1/waiting-queue")
 public class WaitingQueueController {
